move blob_refcounter
