+++ ydb/core/persqueue/subscriber.h
@@ -8,7 +8,7 @@
 #include <ydb/core/tablet/tablet_counters.h>
 #include <ydb/core/base/appdata.h>
 #include <ydb/core/persqueue/events/internal.h>
-#include <ydb/core/persqueue/blob_refcounter.h>
+#include <ydb/core/persqueue/common/blob_refcounter.h>
 
 namespace NKikimr {
 namespace NPQ {
